fix(audio/controller): validate if exists cover or audio image to delete

+++ app/Http/Controllers/AudioController.php
@@ -112,8 +112,14 @@ class AudioController extends Controller
 
     public function destroy(Request $request, Audio $audio): RedirectResponse
     {
-        Storage::disk($audio->disk)->delete($audio->path);
-        Storage::disk($audio->cover_disk)->delete($audio->cover_path);
+        if ($audio->path) {
+            Storage::disk($audio->disk)->delete($audio->path);
+        }
+
+        if ($audio->cover_path) {
+            Storage::disk($audio->cover_disk)->delete($audio->cover_path);
+        }
+
         $audio->delete();
 
         return redirect()->route('home');
